% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalknn_treatment.R
\name{bootstrapIndexMatrixProcessor}
\alias{bootstrapIndexMatrixProcessor}
\title{Finds the NN in the training set for a observation in the training set with a KNN index matrix}
\usage{
bootstrapIndexMatrixProcessor(
  keys_standard,
  index_matrix,
  keys,
  k_value,
  self_as_NN = FALSE
)
}
\arguments{
\item{keys_standard}{A list of the unbootstrapped outcomes and keys in the training set (list)}

\item{index_matrix}{A matrix of indices of the NN (matrix)}

\item{keys}{A vector of the keys in the training set (integer)}

\item{k_value}{A value for K for the causal KNN (integer)}

\item{self_as_NN}{Boolean flag to use yourself as NN. Defaults to FALSE (logical)}
}
\value{
A vector of KNN estimates (numeric)
}
\description{
Takes an pre-searched index matrix and finds the correct nearest nearest neighbors for each
bootstrap iteration.
In the unbootstrapped run, the keys_standard will be the same as the keys.
In a bootstrapped run, the keys parameter will contained the bootstrapped keys.
}
\keyword{internal}
