% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_optimal_k.R
\name{bootstrapKNNStepProcessor}
\alias{bootstrapKNNStepProcessor}
\title{Finds the causal KNN estimates for a specified sequence of K-values}
\usage{
bootstrapKNNStepProcessor(keys_standard, index_matrix, keys, N_step, K_step)
}
\arguments{
\item{keys_standard}{A list of the un-bootstrapped outcomes and keys (list)}

\item{index_matrix}{A matrix of indices of the NN (matrix)}

\item{keys}{A vector of bootstrapped keys (integer)}

\item{N_step}{The number of steps to take (integer)}

\item{k_step}{The value between steps of K and initial K value (integer)}
}
\value{
A matrix of the KNN estimates where rows in NN and cols (matrix)
}
\description{
Helper function for knn_optimal_k.
Takes an pre-searched index matrix and finds the correct nearest nearest neighbors for a sequence
of K's.
Does not include an observation as it own NN
Can be used for bootstrapped data - see the bootstrap vignette
}
\keyword{internal}
