% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casualknn_tep.R
\name{causalKNN_TEP}
\alias{causalKNN_TEP}
\title{Performs Causal KNN Regression and Treatment Effect Projection}
\usage{
causalKNN_TEP(
  DF,
  W,
  Y,
  key,
  DF_test,
  key_test,
  threads = 1L,
  knn_index_mat_options = list(k = floor(sqrt(nrow(DF)))),
  knn_optimal_k_options = list(N_step = ifelse(knn_index_mat_options$k > 300L,
    floor(knn_index_mat_options$k/25), 14L), K_step = ifelse(knn_index_mat_options$k >
    300L, 25L, floor(knn_index_mat_options$k/14))),
  tep_e_net_options = list(),
  verbose = FALSE
)
}
\arguments{
\item{DF}{A data frame of the features                                                                        (data.frame)}

\item{W}{A vector of the treatment indicator (1/0 coded)                                                      (integer)}

\item{Y}{A vector of the outcome values                                                                       (numeric)}

\item{key}{A vector of the indices                                                                            (integer)}

\item{DF_test}{A data frame of the features in the test set                                                   (data.frame)}

\item{key_test}{A vector of the indices of the test sample                                                    (integer)}

\item{threads}{The value of number of threads to use                                                          (integer)}

\item{knn_index_mat_options}{A list of parameters to pass into \code{knn_index_mat}                           (list)}

\item{knn_optimal_k_options}{A list of parameters to pass into \code{knn_optimal_k}                           (list)}

\item{tep_e_net_options}{A list of parameters to pass into \code{tep_projection_e_net}             (list)}

\item{verbose}{A boolean flag for verbose output                                                              (logical)}
}
\value{
A list containing function results (list)
}
\description{
Wraps the causal KNN Regression and the Treatment Effect Projection into one function. The
treatment effect projection step is performed with a Elastic-Net
Sequentially runs \code{knn_index_mat}, \code{knn_optimal_k},
\code{causalknn_treatment_effect},  and \code{tep_projection_e_net} functions and returns a list
of treatment results and optimal parameters. Since all of the important parameters are returned,
the user can re-run certain functions easily. (i.e. re-running \code{causalknn_treatment_effect} with
a larger K value).
}
\details{
This off-the-shelf implementation does not implement bootstrapping for finding the optimal K
value and for the treatment effect projection.
}
