% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalknn_treatment.R
\name{causalknn_treatment_effect}
\alias{causalknn_treatment_effect}
\title{Treatment effect estimation with causal KNN for K values.}
\usage{
causalknn_treatment_effect(
  W,
  Y,
  key,
  K,
  knn_index_list,
  key_test = NULL,
  return_outcomes = FALSE,
  bootstrap_keys = NULL,
  bootstrap_keys_test = NULL
)
}
\arguments{
\item{W}{A vector of the treatment indicator (1/0 coded) in the training set    (integer)}

\item{Y}{A vector of the outcome values in the training set                     (numeric)}

\item{key}{A vector of the indices of the training set                          (integer)}

\item{K}{Value(s) for K                                                         (integer)}

\item{knn_index_list}{A list of the KNN index matrices provided by the knn_index_mat function (list)}

\item{key_test}{A vector of the indices of the test set (Defaults to NULL)      (integer)}

\item{return_outcomes}{Whether to return predicted potential (treated/untreated) outcomes  (logical)}

\item{bootstrap_keys}{A vector of the bootstrapped keys for the training set (default to NULL for no bootstrap case)  (integer)}

\item{bootstrap_keys_test}{A vector of the bootstrapped keys for the test set (default to NULL for no bootstrap case)  (integer)}
}
\value{
A data.frame of the predicted treatment effect, treatment outcome, and untreated outcome by key (data.frame)
}
\description{
In separate \code{K} value case, supply a vector of two elements to the \code{K} parameter. The first
element of vector will be \code{K} for untreated nearest neighbors, and the second elements will be
the K for treated nearest neighbors. Supply a single value for K otherwise.
}
\details{
Supply a \code{key_test} if looking to estimate treatment effects for the validation set. Otherwise,
the default value of NULL will instruct the function to estimate treatment effects for the training set.

Allows for bootstrapped case - see the bootstrap vignette.
}
