% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumWeightedSumStep}
\alias{cumWeightedSumStep}
\title{cumWeightedSumStep}
\usage{
cumWeightedSumStep(Index, y, multiplicity, data, N_steps, K_step)
}
\arguments{
\item{Index}{The matrix of nearest neighbor indices}

\item{y}{A vector of the non-bootstrapped outcomes}

\item{multiplicity}{A vector of the multiplicity for each nearest neighbors observation}

\item{data}{Vector of the bootstrapped indices}

\item{N_steps}{A integer for the number of steps of K_step to take}

\item{K_step}{A integer for the value of steps of K and the initial K value}
}
\value{
A list of the weighted sum and number of NN
}
\description{
This function computes the cumulative weighted sum over K nearest neighbors for a sequence of K values
}
\keyword{internal}
