% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_index_mat.R
\name{knn_index_dist_cov_mahalanobis}
\alias{knn_index_dist_cov_mahalanobis}
\title{Wraps the \code{knn_index_dist_rcpp} function.}
\usage{
knn_index_dist_cov_mahalanobis(
  TRAIN_DATA,
  k,
  TEST_DATA = NULL,
  threads = 1L,
  COV_MAT = NULL
)
}
\arguments{
\item{TRAIN_DATA}{Matrix of the "training" data (matrix)}

\item{k}{Number of NN to find (integer)}

\item{TEST_DATA}{Matrix of the "test" data (matrix)}

\item{threads}{Number of threads to use (integer)}

\item{COV_MAT}{User provided covariance matric for mahalanobis distance (matrix)}
}
\value{
A list containing matricies of the nearest neighbors indicies and distances (list)
}
\description{
General change from the original package's file is that the mahalanobis distance now allows
for a custom co-variance matrix or substitute to be used.
Still supports many other distance metrics such as euclidian or chebyshev. See the original
"KernelKnn" R Package for more details.
}
\details{
Returns a list of NN indicies (matrix) and distances (matrix). For both matricies, the columns
equal the NN and the Rows are the observation.
}
\keyword{internal}
