% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect_projection.R
\name{tep_enet}
\alias{tep_enet}
\title{Performs the TEP step for the Elastic-Net family of models}
\usage{
tep_enet(
  training_X,
  training_TE,
  test_X,
  alpha = 1,
  alpha_seq = seq(0, 1, by = 0.01),
  threads = 1L,
  cv_folds = 10L,
  est_seed = 1234L,
  ...
)
}
\arguments{
\item{training_X}{A data frame of the training set features (data.frame)}

\item{training_TE}{A vector of the individually estimated treatment effects (numeric)}

\item{test_X}{A data frame of the test set features (data.frame)}

\item{alpha}{The alpha value for the elastic net (numeric)}

\item{alpha_seq}{The a vector alpha value to consider (numeric)}

\item{threads}{The value of number of threads to use (integer)}

\item{cv_folds}{The number of folds for the cross-validation step (integer)}

\item{est_seed}{A value for CV estimation seed (integer)}

\item{...}{Additional parameters to be passed to \code{glmnet::cv.glmnet}}
}
\value{
A vector of the TEP projection values for the test set (numeric)
}
\description{
We search over a grid of values for the elastic-net to find the optimal
alpha parameter. Set alpha to NULL to find the optimal alpha value.
The Elastic-Net family of regressions is estimated with `glmnet` package.
Parallelization is done with the `parallel` package.
}
